+++ helpers/file_reader.h
@@ -3,4 +3,4 @@
 
 void file_to_array(char *fileName, char *result[]);
 void file_to_long_long_int_array(char* file_name, long long int result[]);
-void file_to_int_array(char* file_name, long long int result[]);
+void file_to_int_array(char* file_name, int result[]);
